requirements.txt
bidict==0.22.1
blinker==1.6.2
click==8.1.3
dnspython==2.3.0
eventlet==0.33.3
Flask==2.3.2
Flask-SocketIO==5.3.4
greenlet==2.0.2
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
pyserial==3.5
python-engineio==4.4.1
python-socketio==5.8.0
six==1.16.0
Werkzeug==2.3.4
